Add buttons for each choice and call function getRoundResult
Adding three buttons for each rock, paper, scissors. Each button
have an ID to be targeted. When the button is clicked, it will
call function getRoundResult that also calls getComputerChoice,
getPlayerChoice, and playRound function. In order to get the
player choice, I change the function to return the value of
 each button.