Show the score and both player selection.
On index.html add more div for results section so it can show
the player selection and the results. Also add score section
to keep track of the score.

On script.js modify getRoundResult function so it shows
result only for one round. Modify playGame function to show the
score in its section. Modify getWinner function so it will alert
when once the player reach 5 points.